Promise implementation
A very complete implementation of a Promises/A+ equivalent concurrency
primitive for golang. It is safe for access from multiple goroutines
(even in parallel environments) and supports all the features one
would expect and some it might not need.

Even includes tests! This was written because it was asserted that,
while generally the "communicate don't share" paradigm is the right
one, for the purposes of parallelizing large data processing steps it
is likely not. Communication for data-intensive workflows becomes the
bottleneck.