Incomplete promises now block when Get() is called
To allow the caller to effectively wait for a promise chain to
complete, promises have been modified to use a waitgroup to allow
callers to Get() to stall while they are awaiting the result of the
promise chain to be resolved.

Each promise has it's own waitgroup, and as a result, the waitgroup is
marked Done as soon as computation has executed, but before dependency
resolution. This implementation detail should be consistent with the
interface's possible uses and should have no side-effects.

The tests were modified to encourage Get() to fail, but comprehensive
tests were not written.