CompletablePromise.Reject() now captures cause
In what was apparently simply an implementation oversight,
CompletablePromise was not actually capturing the cause when
`Reject(cause)` was called. This would result in subsequent calls to
`Catch()` (which thereafter return a pure rejected promise) not
actually capturing their value.

Further, Combined promises were not properly being rejected when one
of the two promises being bound would fail. The combination of these
two factors lead to a lengthy debugging process — and in the midst of
it, a race condition was discovered on `CompletablePromise.Reject` in
which the `completed` flag would be set to true non-last in a locked
routine. This lead (at times) to the promise being momentarily in an
invalid state in which the to-be rejected promise would appear to
simply be completed successfully, but have no value.

These have been corrected and a simple test (which failed before these
changes) has been added to prove that the promise when rejected always
has a cause.