Complete() and Then() are verifiably reentrant
A deadlock was discovered in the case where `Then()` is called in the
completion chain on a promise which is presently being satisified with
`Complete()`. This has the opportunity to occur naturally in some
software with very complex workflows.

Included a test which verifiably forces this deadlock, which was used
to validate the fix, as well as a more complex test which simply
hammers a single promise concurrently in many goroutines to ensure
that it behaves consistently.

Resolves: #1